Fix to the file lib code to cast opaque values to ints in a more cross-platform manor.

+++ src/functionclose.cpp
@@ -20,7 +20,7 @@ Bu::String FunctionClose::getName() const
 
 Variable FunctionClose::call( Variable &input, VarList )
 {
-	FileMgr::getInstance().close( (int)input.getOpaque() );
+	FileMgr::getInstance().close( (ptrdiff_t)input.getOpaque() );
 	return Variable();
 }
